@@ -54,10 +54,10 @@
     "get-youtube-id": "^1.0.1",
     "lodash": "^4.17.15",
     "ol": "6.15.1",
-    "react": "~17.0.2",
+    "react": "~18.2.0",
     "react-bootstrap": "~2.5.0",
     "react-data-table-component": "^7.0.0",
-    "react-dom": "~17.0.2",
+    "react-dom": "~18.2.0",
     "react-pdf": "^5.3.0",
     "react-player": "^2.2.0",
     "react-select": "^5.0.0",
